split generated code to options.d.ts
