+++ util/gen-typescript-decl.ts
@@ -18,21 +18,12 @@ import * as libtidy from '../';
  * Main entrypoint: generate and update ../src/index.d.ts
  */
 function main() {
-  const dts = path.join(__dirname, ...'../src/index.d.ts'.split('/'));
-  const oldContent = fs.readFileSync(dts, "utf-8");
-  const header = oldContent.split('// START-GENERATED-NAMESPACE')[0];
-  var newContent = header + `// START-GENERATED-NAMESPACE
-// The rest of this file is generated.
-` + generate();
+  const dts = path.join(__dirname, ...'../src/options.d.ts'.split('/'));
+  const content = generate();
 
-  if (newContent === oldContent) {
-    console.log(`no need to update ${dts}`);
-    process.exit(0);
-  } else {
-    fs.writeFileSync(dts, newContent);
-    console.log(`successfully updated ${dts}`);
-    process.exit(0);
-  }
+  fs.writeFileSync(dts, content);
+  console.log(`successfully generated ${dts}`);
+  process.exit(0);
 }
 
 main();
@@ -48,7 +39,7 @@ function createConstUnion(values: (number | string)[]) {
 }
 
 /**
- * format lines generated by dts-dom
+ * re-indent lines generated by dts-dom
  * @param lines lines to format
  */
 function format(lines: string[]) {
@@ -217,7 +208,7 @@ function generate(): string {
   const options = doc.getOptionList();
 
   const nm = dom.create.namespace("Generated");
-  nm.flags = dom.DeclarationFlags.None;
+  nm.flags = dom.DeclarationFlags.Export;
   nm.jsDocComment = `
 Type for libtidy options
 @generated with /util/gen-typescript-decl.ts
